backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
whitenoise>=6.6,<7.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
openpyxl>=3.1,<4.0
WeasyPrint>=62.0,<63.0
